Refactor structure to enable code splitting
